% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linreg Class calculates Multiple Linear Regression Model For Data frame.}
\value{
results from the linreg function
}
\description{
Linreg Class calculates Multiple Linear Regression Model For Data frame.
}
\section{Fields}{

\describe{
\item{\code{formula}}{Dependent and independent columns in data frame}

\item{\code{data}}{Data frame for linear models}

\item{\code{coeffs}}{Estimators matrix converted to vector}

\item{\code{pred_val}}{(matrix). Estimated dependent values}

\item{\code{e}}{(matrix) Error calculated from actual minus calculated dependent variable}

\item{\code{deg_f}}{degrees of freedom}

\item{\code{var_e}}{(matrix) Residual variance}

\item{\code{var_coeffs}}{variance of the regression coefficients}

\item{\code{t_value}}{(matrix) T value}

\item{\code{df_name}}{(character) Data Frame name}
}}


