% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizing_flight_delays.R
\name{Visualizing_flight_delay}
\alias{Visualizing_flight_delay}
\title{Visualizing the mean delays of the airports}
\usage{
Visualizing_flight_delay()
}
\value{
plot of the mean delays of different airports by longititude and latitude
}
\description{
Visualizing the mean delays of the airports
}
