% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearregression.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Linear Regression Model}
\format{
An object of class \code{"linreg"}.
}
\description{
A reference class for multiple linear regression models.
}
\section{Fields}{

\describe{
\item{\code{formula}}{A formula object specifying the regression model.}

\item{\code{data}}{A data frame containing the data.}

\item{\code{X}}{The design matrix of independent variables.}

\item{\code{y}}{The dependent variable.}

\item{\code{lm_model}}{The fitted linear regression model.}
}}


\section{Methods}{

\describe{
  \method{print}{linreg}: Print coefficients and coefficient names.
  \method{plot}{linreg}: Plot residuals vs. fitted values and standardized residuals vs. fitted values.
  \method{resid}{linreg}: Return the vector of residuals.
  \method{pred}{linreg}: Return the predicted values.
  \method{coef}{linreg}: Return the coefficients as a named vector.
  \method{summary}{linreg}: Return a summary of the linear regression model.
}
}

\examples{
# Create a linreg object
lin_reg_model <- linreg(formula = Petal.Length ~ Species, data = iris)
# Print coefficients
print(lin_reg_model)
# Plot residuals
plot(lin_reg_model)
# Get residuals
residuals <- resid(lin_reg_model)
# Get predicted values
predictions <- pred(lin_reg_model)
# Get coefficients
coefficients <- coef(lin_reg_model)
# Summary of the model
summary(lin_reg_model)
}
